Rename Sanitizer back to Formatter.
