+++ index.php
@@ -25,11 +25,11 @@ if (F::exists(__DIR__ . '/vendor/autoload.php') === false) {
         'sgkirby\\commentions\\Endpoint' => 'classes/Endpoint.php',
         'sgkirby\\commentions\\Frontend' => 'classes/Frontend.php',
         'sgkirby\\commentions\\Migration' => 'classes/Migration.php',
-        'sgkirby\\commentions\\Sanitizer' => 'classes/Sanitizer.php',
-        'sgkirby\\commentions\\Sanitizer\\CacheAdapter' => 'classes/Sanitizer/CacheAdapter.php',
-        'sgkirby\\commentions\\Sanitizer\\CodeClassAttrDef' => 'classes/Sanitizer/CodeClassAttrDef.php',
-        'sgkirby\\commentions\\Sanitizer\\LinkTransformer' => 'classes/Sanitizer/LinkTransformer.php',
-        'sgkirby\\commentions\\Sanitizer\\RemoveEmptyLinksInjector' => 'classes/Sanitizer/RemoveEmptyLinksInjector.php',
+        'sgkirby\\commentions\\Formatter' => 'classes/Formatter.php',
+        'sgkirby\\commentions\\Formatter\\CacheAdapter' => 'classes/Formatter/CacheAdapter.php',
+        'sgkirby\\commentions\\Formatter\\CodeClassAttrDef' => 'classes/Formatter/CodeClassAttrDef.php',
+        'sgkirby\\commentions\\Formatter\\LinkTransformer' => 'classes/Formatter/LinkTransformer.php',
+        'sgkirby\\commentions\\Formatter\\RemoveEmptyLinksInjector' => 'classes/Formatter/RemoveEmptyLinksInjector.php',
         'sgkirby\\commentions\\Storage' => 'classes/Storage.php',
         'sgkirby\\commentions\\Structure' => 'classes/Structure.php',
     ], __DIR__);
