+++ sections/commentions.php
@@ -68,7 +68,7 @@ return [
                 ];
             }
 
-            if (class_exists('Masterminds\\HTML5') === false || Sanitizer::available() === false) {
+            if (class_exists('Masterminds\\HTML5') === false || Formatter::available() === false) {
                 $errors[] = [
                     'id' => 'missing-dependencies',
                     'message' => t('commentions.section.error.missing-dependencies'),
